requirements.txt
gdown==4.0.1
numpy==1.20.0
pandas==1.3.2
Pillow==10.3.0
PyYAML==6.0.1
scikit-learn==0.24.1
scipy==1.7.0
timm==0.8.6.dev0
tqdm==4.66.2
transformers==4.24.0
numba==0.56.4
llvmlite==0.39.1
umap-learn==0.5.7